nav2_bridge: map switch 系を別 TU へ分離 (#345)
